# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fauser Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List study materials
      parameters:
        - name: classId
          in: query
          schema:
            type: ["integer", "null"]
        - name: subjectId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
    post:
      operationId: createMaterial
      tags: [materials]
      summary: Upload a material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialInput"
      responses:
        "201":
          description: Created material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
  /materials/{id}:
    delete:
      operationId: deleteMaterial
      tags: [materials]
      summary: Delete a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /polls:
    get:
      operationId: listPolls
      tags: [polls]
      summary: List polls with vote counts
      responses:
        "200":
          description: Polls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PollDetail"
    post:
      operationId: createPoll
      tags: [polls]
      summary: Create a poll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PollInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollDetail"
  /polls/{id}/vote:
    post:
      operationId: votePoll
      tags: [polls]
      summary: Vote on a poll
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PollVoteInput"
      responses:
        "200":
          description: Updated poll
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollDetail"
  /polls/{id}:
    patch:
      operationId: updatePollStatus
      tags: [polls]
      summary: Close or update poll
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PollStatusInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollDetail"
...
  /tutoring:
    get:
      operationId: listTutoringPosts
      tags: [tutoring]
      summary: List tutoring posts
      parameters:
        - in: query
          name: subjectId
          schema:
            type: integer
        - in: query
          name: type
          schema:
            type: string
            enum: [offre, cerca]
      responses:
        "200":
          description: Tutoring posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TutoringPost"
    post:
      operationId: createTutoringPost
      tags: [tutoring]
      summary: Create a tutoring post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TutoringPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutoringPost"
  /tutoring/{id}:
    patch:
      operationId: updateTutoringPost
      tags: [tutoring]
      summary: Update tutoring post status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TutoringPostStatusInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutoringPost"
    delete:
      operationId: deleteTutoringPost
      tags: [tutoring]
      summary: Delete a tutoring post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Material:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        subjectId:
          type: integer
        subjectName:
          type: string
        classId:
          type: integer
        className:
          type: string
        fileUrl:
          type: string
        fileType:
          type: ["string", "null"]
        fileSize:
          type: ["integer", "null"]
        uploadedById:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - title
        - subjectId
        - subjectName
        - classId
        - className
        - fileUrl
        - uploadedById
        - createdAt
    MaterialInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        subjectId:
          type: integer
        classId:
          type: integer
        fileUrl:
          type: string
        fileType:
          type: ["string", "null"]
        fileSize:
          type: ["integer", "null"]
      required:
        - title
        - subjectId
        - classId
        - fileUrl
...
    PollDetail:
      type: object
      properties:
        id:
          type: integer
        question:
          type: string
        authorId:
          type: integer
        authorName:
          type: string
        classId:
          type: integer
          nullable: true
        status:
          type: string
        expiresAt:
          type: string
          nullable: true
        createdAt:
          type: string
        options:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              text:
                type: string
              voteCount:
                type: integer
            required:
              - id
              - text
              - voteCount
        totalVotes:
          type: integer
        myVoteOptionId:
          type: integer
          nullable: true
      required:
        - id
        - question
        - authorId
        - authorName
        - status
        - createdAt
        - options
        - totalVotes
    PollInput:
      type: object
      properties:
        question:
          type: string
        options:
          type: array
          items:
            type: string
        classId:
          type: integer
        expiresAt:
          type: string
      required:
        - question
        - options
    PollVoteInput:
      type: object
      properties:
        optionId:
          type: integer
      required:
        - optionId
    PollStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [active, closed]
      required:
        - status
...
    TutoringPost:
      type: object
      properties:
        id:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        subjectId:
          type: integer
        subjectName:
          type: string
        type:
          type: string
        description:
          type: string
        status:
          type: string
        classId:
          type: integer
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - authorId
        - authorName
        - subjectId
        - subjectName
        - type
        - description
        - status
        - createdAt
    TutoringPostStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [active, closed]
      required:
        - status
    TutoringPostInput:
      type: object
      properties:
        subjectId:
          type: integer
        type:
          type: string
        description:
          type: string
        classId:
          type: integer
      required:
        - subjectId
        - type
        - description
...