# lib/api-zod/src/generated/types/materialInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fauser Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface MaterialInput {
  title: string;
  /** @nullable */
  description?: string | null;
  subjectId: number;
  classId: number;
  fileUrl: string;
  /** @nullable */
  fileType?: string | null;
  /** @nullable */
  fileSize?: number | null;
}